примитивы, equal

+++ RVM/Functions/GC.cpp
@@ -1,8 +1,9 @@
 #include "GC.h"
+#include "../Core/FieldObject.h"
 
 void GC()
 {
-	for (int i = 0; i < GlobalVariables::heap.GetSize(); ++i)
+	for (size_t i = 0; i < GlobalVariables::heap.GetSize(); ++i)
 	{
 		if (GlobalVariables::heap[i] != nullptr)
 		if (GlobalVariables::heap[i]->GetAmountUsers() <= 0)
